+++ src/components/Layout/Backoffice/index.ts
@@ -1,0 +1,4 @@
+import Backoffice from "./Backoffice";
+
+export default Backoffice;
+export * from "./Backoffice";