+++ src/components/Surface/AppBar/index.ts
@@ -1,0 +1,4 @@
+import AppBar from "./AppBar";
+
+export default AppBar;
+export * from "./AppBar";