+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # [Versions](https://github.com/Tracktor/design-system/releases)
 
-## v1.58.1
-- **[feat]** - add `TimeLine` component
+## v1.59.0
+- **[feat]** - add `Backoffice` component
+- **[feat]** - add `Header` component
+- **[feat]** - New variant `shape-background` for `Logo` component
+- **[chore]** - Improve module augmentation declaration